requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.0
ultralytics==8.0.145  # YOLO
face-recognition==1.3.0
scikit-learn==1.3.0  # For clustering
moviepy==1.0.3  # For video processing
requests==2.31.0
python-jose==3.3.0
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6